Add Auth Jwt StartUp registration methods, Swagger registration
