+++ Spark.Templates/working/Spark.Templates.csproj
@@ -52,4 +52,8 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="templates\Spark.Templates.Api\Application\Database\Migrations\" />
+  </ItemGroup>
+
 </Project>
