+++ Spark.Templates/working/templates/Spark.Templates.Api/Application/Services/Auth/AuthService.cs
@@ -1,18 +1,17 @@
-﻿using Spark.Templates.Api.Application.Services.Auth;
-using Spark.Library.Auth;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using System.Security.Claims;
+﻿using Microsoft.IdentityModel.Tokens;
 using Spark.Templates.Api.Application.Models;
 using System.Globalization;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
 
 namespace Spark.Templates.Api.Application.Services.Auth
 {
-    public class AuthService
+	public class AuthService
 	{
 		private readonly UsersService _usersService;
 		private readonly RolesService _rolesService;
-		private readonly IConfiguration _config;
+		private readonly IConfiguration _configuration;
 		private readonly IHttpContextAccessor _contextAccessor;
 		private HttpContext _context;
 
@@ -36,7 +35,7 @@ namespace Spark.Templates.Api.Application.Services.Auth
 		public AuthService(RolesService rolesService, IConfiguration config, IHttpContextAccessor contextAccessor, UsersService usersService)
 		{
 			_rolesService = rolesService;
-			_config = config;
+			_configuration = config;
 			_contextAccessor = contextAccessor;
 			_usersService = usersService;
 		}
@@ -66,39 +65,58 @@ namespace Spark.Templates.Api.Application.Services.Auth
 			return userId;
 		}
 
-		public async Task RefreshSignIn(User user)
+		public async Task<string> CreateJwtToken(User user)
 		{
-			await Context.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
-
-			var cookieClaims = await this.CreateCookieClaims(user);
-			var loginCookieExpirationDays = _config.GetValue("Spark:Auth:CookieExpirationDays", 5);
-			await Context.SignInAsync(
-				CookieAuthenticationDefaults.AuthenticationScheme,
-				cookieClaims,
-				new AuthenticationProperties
-				{
-					IsPersistent = true,
-					IssuedUtc = DateTimeOffset.UtcNow,
-					ExpiresUtc = DateTimeOffset.UtcNow.AddDays(loginCookieExpirationDays)
-				});
+			var tokenExpirationDays = _configuration.GetValue("Spark:Jwt:ExpirationDays", 5);
+			var expiration = DateTime.UtcNow.AddDays(tokenExpirationDays);
+			var token = CreateJwtSecurityToken(
+				await CreateJwtClaimsAsync(user),
+				CreateJwtSigningCredentials(),
+				expiration
+			);
+			var tokenHandler = new JwtSecurityTokenHandler();
+			return tokenHandler.WriteToken(token);
 		}
 
-		public async Task<ClaimsPrincipal> CreateCookieClaims(User user)
-        {
-            var identity = new ClaimsIdentity(CookieAuthenticationDefaults.AuthenticationScheme);
-            identity.AddClaim(new Claim(ClaimTypes.NameIdentifier, user.Id.ToString(CultureInfo.InvariantCulture)));
-            identity.AddClaim(new Claim(ClaimTypes.Email, user.Email));
-            identity.AddClaim(new Claim(ClaimTypes.Name, user.Name));
-            identity.AddClaim(new Claim(ClaimTypes.UserData, user.Id.ToString(CultureInfo.InvariantCulture)));
+		private JwtSecurityToken CreateJwtSecurityToken(List<Claim> claims, SigningCredentials credentials,
+		DateTime expiration) =>
+		new(
+				_configuration.GetValue("Spark:Jwt:Issuer", "https://spark-framework.net"),
+				_configuration.GetValue("Spark:Jwt:Audience", "https://spark-framework.net"),
+				claims,
+				expires: expiration,
+				signingCredentials: credentials
+			);
 
-            // add roles
-            var roles = await _rolesService.FindUserRolesAsync(user.Id);
-            foreach (var role in roles)
-            {
-                identity.AddClaim(new Claim(ClaimTypes.Role, role.Name));
-            }
+		private async Task<List<Claim>> CreateJwtClaimsAsync(User user)
+		{
+			var claims = new List<Claim>
+				{
+					new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
+					new Claim(JwtRegisteredClaimNames.Iat, DateTime.UtcNow.ToString(CultureInfo.InvariantCulture)),
+					new Claim(ClaimTypes.NameIdentifier, user.Id.ToString(CultureInfo.InvariantCulture)),
+					new Claim(ClaimTypes.Email, user.Email),
+					new Claim(ClaimTypes.Name, user.Name),
+					new Claim(ClaimTypes.UserData, user.Id.ToString(CultureInfo.InvariantCulture))
+				};
 
-            return new ClaimsPrincipal(identity);
-        }
-    }
+			// add roles
+			var roles = await _rolesService.FindUserRolesAsync(user.Id);
+			foreach (var role in roles)
+			{
+				claims.Add(new Claim(ClaimTypes.Role, role.Name));
+			}
+
+			return claims;
+		}
+		private SigningCredentials CreateJwtSigningCredentials()
+		{
+			return new SigningCredentials(
+				new SymmetricSecurityKey(
+					Encoding.UTF8.GetBytes(_configuration.GetValue("Spark:Jwt:Key", "SomthingSecret!"))
+				),
+				SecurityAlgorithms.HmacSha256
+			);
+		}
+	}
 }
