--- Spark.Templates/working/templates/Spark.Templates.Api/Application/ViewModels/User.cs
@@ -1,11 +1,0 @@
-﻿namespace Spark.Templates.Api.Application.ViewModels
-{
-	public class User
-	{
-		public string Name { get; set; }
-
-		public string Email { get; set; }
-
-		public DateTime? EmailVerifiedAt { get; set; }
-	}
-}
